use new sidebar component
