+++ src/components/Sidebar/sidebar.scss
@@ -10,15 +10,8 @@
       margin-top: 10px;
     }
   }
-  .line {
-    height: 0;
-    border: 1px solid #475467;
-    width: 85%;
-    padding: 0 !important;
-    margin: 0px auto;
-  }
   .center {
-    height: 68vh !important;
+    height: 80% !important;
     overflow: auto;
 
     ul {
@@ -86,10 +79,7 @@
       text-align: center;
     }
   }
-  .link {
-    text-decoration: none;
-    color: inherit;
-  }
+
   .active {
     font-weight: 800;
   }
@@ -99,8 +89,10 @@
   }
 }
 .center {
-  height: 68vh;
-  overflow: auto; /* Keep the scrollbar visible */
+  height: 50%;
+  overflow: auto;
+  margin-top: 10px;
+  margin-bottom: 10px; /* Keep the scrollbar visible */
 }
 
 .center::-webkit-scrollbar {
@@ -110,12 +102,55 @@
 .center::-webkit-scrollbar-thumb {
   background-color: #475467; /* Set the color of the scrollbar thumb */
 }
-.main-container-sidebar{
+
+.line {
+  height: 0;
+  border: 1px solid #475467;
+  width: 85%;
+  padding: 0 !important;
+  margin: 0px auto;
+}
+.main-container-sidebar {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
+  height: 92%;
+}
+.link {
+  text-decoration: none;
+  color: inherit;
+}
+.button-item {
+  height: 20px;
+  color: white;
+  font-size: 16px;
+  border-radius: 6px;
+  text-align: left;
+  display: flex;
+  gap: 10px;
+  padding: 12px;
+  cursor: pointer;
+  align-items: center;
+  //   background-color: #8477da;
+  .active {
+    background-color: #8477da;
+    border-radius: 6px;
+  }
+}
+.button-item:hover {
+  background-color: #8477da;
 }
 .active {
   background-color: #8477da;
   border-radius: 6px;
+}
+.subtitles {
+  color: white;
+  font-size: 20px !important;
+  font-weight: bold !important;
+}
+@media screen and (max-width: 1600px) {
+  .main-container-sidebar {
+    height: 90%;
+  }
 }
